+++ packages/terra-button/src/_constants.js
@@ -1,0 +1,21 @@
+const ButtonVariants = {
+  NEUTRAL: 'neutral',
+  EMPHASIS: 'emphasis',
+  GHOST: 'ghost',
+  'DE-EMPHASIS': 'de-emphasis',
+  ACTION: 'action',
+  UTILITY: 'utility',
+};
+
+const ButtonTypes = {
+  BUTTON: 'button',
+  SUBMIT: 'submit',
+  RESET: 'reset',
+};
+
+const IconTypes = {
+  DECORATIVE: 'decorative',
+  INFORMATIVE: 'informative',
+};
+
+export { ButtonVariants, ButtonTypes, IconTypes };