+++ packages/terra-core-docs/src/terra-dev-site/test/status-view/StatusViewCustom.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import IconDiamond from 'terra-icon/lib/icon/IconDiamond';
+import IconDiamond from 'terra-icon/lib/icon/decorative/IconDiamond';
 import StatusView from 'terra-status-view';
 
 const StatusViewButtons = [